Restrict exports in server actions
