+++ src/app/groups/GroupAction.ts
@@ -1,7 +1,6 @@
 'use server';
 
 import { uniqueArray } from '../../common/arrayUtils';
-import { UserRecordExtended } from '../../common/users';
 import {
   GROUP_COLLECTION_ID,
   USER_COLLECTION_ID,
@@ -12,21 +11,7 @@ import {
   setCustomClaimsForUser,
 } from '../api/users/[uid]/updateUser';
 import { actionAdminRequired, actionUserRequired } from '../auth';
-
-export interface Group {
-  id?: string;
-  name: string;
-  description?: string;
-}
-
-async function getGroups(): Promise<Group[]> {
-  const groupDocs = (
-    await firestore.collection(GROUP_COLLECTION_ID).orderBy('name', 'asc').get()
-  ).docs;
-  return groupDocs.map(
-    (g) => ({ ...g.data(), name: g.data().name || '', id: g.id } as Group)
-  );
-}
+import { getGroups, getMyGroups, Group } from './groupHelpers';
 
 export async function getGroupsAction(): Promise<Group[]> {
   await actionUserRequired();
@@ -119,19 +104,6 @@ export async function updateGroupAction(group: Group, assigendUsers: string[]) {
     )
   );
   return groupId;
-}
-
-async function getMyGroups(userId: string): Promise<Group[]> {
-  const allGropus = await getGroups();
-  const myGroupIds =
-    (
-      (
-        await firestore.collection(USER_COLLECTION_ID).doc(userId).get()
-      ).data() as UserRecordExtended
-    ).groups || [];
-  return allGropus
-    .filter((g) => g.id && myGroupIds.includes(g.id))
-    .sort((a, b) => a.name.localeCompare(b.name));
 }
 
 export async function getMyGroupsFromServer(): Promise<Group[]> {
