+++ src/app/groups/GroupDialog.tsx
@@ -1,20 +1,20 @@
 import Button from '@mui/material/Button';
+import Checkbox from '@mui/material/Checkbox';
 import Dialog from '@mui/material/Dialog';
 import DialogActions from '@mui/material/DialogActions';
 import DialogContent from '@mui/material/DialogContent';
 import DialogContentText from '@mui/material/DialogContentText';
 import DialogTitle from '@mui/material/DialogTitle';
-import Select, { SelectChangeEvent } from '@mui/material/Select';
-import TextField from '@mui/material/TextField';
-import React, { useEffect, useMemo, useState } from 'react';
-import { Group } from './GroupAction';
-import { UserRecordExtended } from '../../common/users';
 import FormControl from '@mui/material/FormControl';
 import InputLabel from '@mui/material/InputLabel';
-import MenuItem from '@mui/material/MenuItem';
-import Checkbox from '@mui/material/Checkbox';
 import ListItemText from '@mui/material/ListItemText';
+import MenuItem from '@mui/material/MenuItem';
 import OutlinedInput from '@mui/material/OutlinedInput';
+import Select, { SelectChangeEvent } from '@mui/material/Select';
+import TextField from '@mui/material/TextField';
+import React, { useEffect, useState } from 'react';
+import { UserRecordExtended } from '../../common/users';
+import { Group } from './groupHelpers';
 
 export interface GroupDialoggOptions {
   onClose: (item?: Group, assigendUsers?: string[]) => void;
